+++ drivers/gpu/msm/kgsl_pwrctrl.c
@@ -163,7 +163,7 @@ void kgsl_pwrctrl_pwrlevel_change(struct kgsl_device *device,
 
 EXPORT_SYMBOL(kgsl_pwrctrl_pwrlevel_change);
 
-static int kgsl_pwrctrl_thermal_pwrlevel_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_thermal_pwrlevel_store(struct device *dev,
 					 struct device_attribute *attr,
 					 const char *buf, size_t count)
 {
@@ -203,7 +203,7 @@ static int kgsl_pwrctrl_thermal_pwrlevel_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_thermal_pwrlevel_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_thermal_pwrlevel_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -216,7 +216,7 @@ static int kgsl_pwrctrl_thermal_pwrlevel_show(struct device *dev,
 	return snprintf(buf, PAGE_SIZE, "%d\n", pwr->thermal_pwrlevel);
 }
 
-static int kgsl_pwrctrl_max_pwrlevel_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_max_pwrlevel_store(struct device *dev,
 					 struct device_attribute *attr,
 					 const char *buf, size_t count)
 {
@@ -257,7 +257,7 @@ static int kgsl_pwrctrl_max_pwrlevel_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_max_pwrlevel_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_max_pwrlevel_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -270,7 +270,7 @@ static int kgsl_pwrctrl_max_pwrlevel_show(struct device *dev,
 	return snprintf(buf, PAGE_SIZE, "%u\n", pwr->max_pwrlevel);
 }
 
-static int kgsl_pwrctrl_min_pwrlevel_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_min_pwrlevel_store(struct device *dev,
 					 struct device_attribute *attr,
 					 const char *buf, size_t count)
 {	struct kgsl_device *device = kgsl_device_from_dev(dev);
@@ -312,7 +312,7 @@ static int kgsl_pwrctrl_min_pwrlevel_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_min_pwrlevel_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_min_pwrlevel_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -324,7 +324,7 @@ static int kgsl_pwrctrl_min_pwrlevel_show(struct device *dev,
 	return snprintf(buf, PAGE_SIZE, "%u\n", pwr->min_pwrlevel);
 }
 
-static int kgsl_pwrctrl_active_pwrlevel_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_active_pwrlevel_show(struct device *dev,
                                         struct device_attribute *attr,
                                         char *buf)
 {
@@ -336,7 +336,7 @@ static int kgsl_pwrctrl_active_pwrlevel_show(struct device *dev,
 	return snprintf(buf, PAGE_SIZE, "%d\n", pwr->active_pwrlevel);
 }
 
-static int kgsl_pwrctrl_num_pwrlevels_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_num_pwrlevels_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -363,7 +363,7 @@ static int _get_nearest_pwrlevel(struct kgsl_pwrctrl *pwr, unsigned int clock)
 	return -ERANGE;
 }
 
-static int kgsl_pwrctrl_max_gpuclk_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_max_gpuclk_store(struct device *dev,
 					 struct device_attribute *attr,
 					 const char *buf, size_t count)
 {
@@ -408,7 +408,7 @@ done:
 	return count;
 }
 
-static int kgsl_pwrctrl_max_gpuclk_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_max_gpuclk_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -422,7 +422,7 @@ static int kgsl_pwrctrl_max_gpuclk_show(struct device *dev,
 			pwr->pwrlevels[pwr->thermal_pwrlevel].gpu_freq);
 }
 
-static int kgsl_pwrctrl_gpuclk_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_gpuclk_store(struct device *dev,
 				     struct device_attribute *attr,
 				     const char *buf, size_t count)
 {
@@ -449,7 +449,7 @@ static int kgsl_pwrctrl_gpuclk_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_gpuclk_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_gpuclk_show(struct device *dev,
 				    struct device_attribute *attr,
 				    char *buf)
 {
@@ -469,7 +469,7 @@ static int kgsl_pwrctrl_gpuclk_show(struct device *dev,
 	return snprintf(buf, PAGE_SIZE, "%lu\n", freq);
 }
 
-static int kgsl_pwrctrl_idle_timer_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_idle_timer_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count)
 {
@@ -496,7 +496,7 @@ static int kgsl_pwrctrl_idle_timer_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_idle_timer_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_idle_timer_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -509,7 +509,7 @@ static int kgsl_pwrctrl_idle_timer_show(struct device *dev,
 		device->pwrctrl.interval_timeout * mul);
 }
 
-static int kgsl_pwrctrl_pmqos_active_latency_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_pmqos_active_latency_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count)
 {
@@ -531,7 +531,7 @@ static int kgsl_pwrctrl_pmqos_active_latency_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_pmqos_active_latency_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_pmqos_active_latency_show(struct device *dev,
 					   struct device_attribute *attr,
 					   char *buf)
 {
@@ -575,7 +575,7 @@ static ssize_t kgsl_pwrctrl_pmqos_wakeup_latency_show(struct device *dev,
 		device->pwrctrl.pm_qos_wakeup_latency);
 }
 
-static int kgsl_pwrctrl_gpubusy_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_gpubusy_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -595,7 +595,7 @@ static int kgsl_pwrctrl_gpubusy_show(struct device *dev,
 	return ret;
 }
 
-static int kgsl_pwrctrl_gpu_available_frequencies_show(
+static ssize_t kgsl_pwrctrl_gpu_available_frequencies_show(
 					struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
@@ -614,7 +614,7 @@ static int kgsl_pwrctrl_gpu_available_frequencies_show(
 	return num_chars;
 }
 
-static int kgsl_pwrctrl_reset_count_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_reset_count_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -645,7 +645,7 @@ static void __force_on(struct kgsl_device *device, int flag, int on)
 	}
 }
 
-static int __force_on_show(struct device *dev,
+static ssize_t __force_on_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf, int flag)
 {
@@ -656,7 +656,7 @@ static int __force_on_show(struct device *dev,
 		test_bit(flag, &device->pwrctrl.ctrl_flags));
 }
 
-static int __force_on_store(struct device *dev,
+static ssize_t __force_on_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count,
 					int flag)
@@ -679,42 +679,42 @@ static int __force_on_store(struct device *dev,
 	return count;
 }
 
-static int kgsl_pwrctrl_force_clk_on_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_clk_on_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
 	return __force_on_show(dev, attr, buf, KGSL_PWRFLAGS_CLK_ON);
 }
 
-static int kgsl_pwrctrl_force_clk_on_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_clk_on_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count)
 {
 	return __force_on_store(dev, attr, buf, count, KGSL_PWRFLAGS_CLK_ON);
 }
 
-static int kgsl_pwrctrl_force_bus_on_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_bus_on_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
 	return __force_on_show(dev, attr, buf, KGSL_PWRFLAGS_AXI_ON);
 }
 
-static int kgsl_pwrctrl_force_bus_on_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_bus_on_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count)
 {
 	return __force_on_store(dev, attr, buf, count, KGSL_PWRFLAGS_AXI_ON);
 }
 
-static int kgsl_pwrctrl_force_rail_on_show(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_rail_on_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
 	return __force_on_show(dev, attr, buf, KGSL_PWRFLAGS_POWER_ON);
 }
 
-static int kgsl_pwrctrl_force_rail_on_store(struct device *dev,
+static ssize_t kgsl_pwrctrl_force_rail_on_store(struct device *dev,
 					struct device_attribute *attr,
 					const char *buf, size_t count)
 {
